refactor: move and rename EditingPanel to TabPanel
