+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Provider } from "react-redux"
 import store from "@/redux/store"
 
 import Canvas from "@/components/Canvas/Canvas"
-import EditingPanel from "@/components/EditingPanel"
+import EditingPanel from "@/components/Tab/TabPanel"
 
 function App() {
   return (
